refactor(macos): split managed image loading
