+++ apps/macos/Sources/OpenClaw/GatewayConnection.swift
@@ -6,7 +6,6 @@ import OpenClawProtocol
 import OSLog
 
 private let gatewayConnectionLogger = Logger(subsystem: "ai.openclaw", category: "gateway.connection")
-private let gatewayManagedImagePathPrefix = "/api/chat/media/outgoing/"
 
 /// Single, shared Gateway websocket connection for the whole app.
 ///
@@ -48,10 +47,10 @@ actor GatewayConnection {
     struct Route: Equatable, Sendable {
         fileprivate let generation: UInt64
         fileprivate let authority: UInt64?
-        fileprivate let url: URL
+        let url: URL
         fileprivate let token: String?
         fileprivate let password: String?
-        fileprivate let tls: GatewayTLSRoute?
+        let tls: GatewayTLSRoute?
         fileprivate let deviceAuthGatewayID: String?
         let activationOwnershipFingerprint: String?
 
@@ -68,7 +67,9 @@ actor GatewayConnection {
     /// One connected Gateway server, not merely an endpoint configuration.
     /// A reconnect at the same URL creates a different lease.
     struct ServerLease: Sendable {
-        fileprivate let route: Route
+        // Managed-image HTTP reuses this captured route from its focused extension file.
+        // Carrying the snapshot forward prevents endpoint or TLS rediscovery after suspension.
+        let route: Route
         fileprivate let socketGeneration: UInt64
         fileprivate let client: GatewayChannelActor
     }
@@ -444,73 +445,6 @@ actor GatewayConnection {
             }
             throw OpenClawChatTransportSendError.notDispatched
         }
-    }
-
-    func loadImageArtifact(
-        sessionKey: String,
-        agentID: String?,
-        artifactId: String,
-        ifCurrentServerLease lease: ServerLease) async throws -> OpenClawChatLoadedImage?
-    {
-        let request = OpenClawChatGatewayRequests.artifactDownload(
-            sessionKey: sessionKey,
-            agentID: agentID,
-            artifactId: artifactId)
-        let responseData = try await self.request(
-            method: request.method,
-            params: request.params,
-            timeoutMs: request.timeoutMs,
-            ifCurrentServerLease: lease)
-        let response = try JSONDecoder().decode(ArtifactsDownloadResult.self, from: responseData)
-        guard let ticketedPath = response.url?.trimmingCharacters(in: .whitespacesAndNewlines),
-              let url = Self.managedImageURL(gatewayURL: lease.route.url, ticketedPath: ticketedPath)
-        else { return nil }
-
-        var urlRequest = URLRequest(url: url)
-        urlRequest.timeoutInterval = 20
-        urlRequest.setValue("image/*", forHTTPHeaderField: "Accept")
-        let tls = lease.route.tls?.params ?? GatewayTLSParams(
-            required: false,
-            expectedFingerprint: nil,
-            allowTOFU: false,
-            storeKey: nil)
-        let session = GatewayTLSPinningSession(params: tls)
-        defer { session.finishTasksAndInvalidate() }
-        let (data, urlResponse) = try await session.data(
-            for: urlRequest,
-            maximumBytes: 12 * 1024 * 1024)
-        guard await self.isCurrentServerLease(lease) else {
-            throw OpenClawChatTransportSendError.notDispatched
-        }
-        guard let http = urlResponse as? HTTPURLResponse,
-              (200..<300).contains(http.statusCode),
-              let mimeType = http.mimeType?.lowercased(),
-              mimeType.hasPrefix("image/")
-        else { return nil }
-        return OpenClawChatLoadedImage(data: data, mimeType: mimeType)
-    }
-
-    private static func managedImageURL(gatewayURL: URL, ticketedPath: String) -> URL? {
-        guard ticketedPath.hasPrefix(gatewayManagedImagePathPrefix),
-              let relative = URLComponents(string: ticketedPath),
-              relative.scheme == nil,
-              relative.host == nil,
-              relative.fragment == nil,
-              relative.queryItems?.contains(where: {
-                  $0.name == "mediaTicket" && $0.value?.isEmpty == false
-              }) == true,
-              var base = URLComponents(url: gatewayURL, resolvingAgainstBaseURL: false),
-              base.host != nil
-        else { return nil }
-        switch base.scheme?.lowercased() {
-        case "wss", "https": base.scheme = "https"
-        case "ws", "http": base.scheme = "http"
-        default: return nil
-        }
-        base.percentEncodedPath = relative.percentEncodedPath
-        base.percentEncodedQuery = relative.percentEncodedQuery
-        base.fragment = nil
-        return base.url
     }
 }
